Move channel reducer
