+++ packages/wallet/src/redux/initializedReducer.ts
@@ -10,7 +10,7 @@ import {
 import { WalletAction, CHANNEL_INITIALIZED } from './actions';
 import { channelInitializationSuccess } from 'magmo-wallet-client/lib/wallet-events';
 import { ethers } from 'ethers';
-import { channelReducer } from './reducers/channels';
+import { channelReducer } from './channel/reducer';
 import { unreachable } from '../utils/reducer-utils';
 import { outboxStateReducer } from './reducer';
